Make protonvpn-cli work on OpenVPN 2.3 and older.
I still have a system or two that is too old to have Python 3.5+ (which
is required for the newer, supported linux-cli ProtonVPN client) and
OpenVPN versions older than 2.4 in which directives like "pull-filter"
are pushed down.

To get around this and saved this for posterity for my own uses I've
employed a rather simple sed hack to strip pull-filter references from
the generated-on-runtime protonvpn_openvpn_config.conf that the
deprecated protonvpn-cli uses.